+++ CHANGELOG.md
@@ -17,7 +17,11 @@ All notable changes to `Kronika` will be documented in this file.
 - `Second::tryOf()` method.
 - `Instant::tryOf()` and `Instant::tryOfValue()` methods.
 
-### Changed
+### Deprecated
+- `precision` argument for `Range\DateTimeRange::of()` and `Range\DateTimeRange::around()`.
+- `precision` argument for `Range\TimeRange::of()`.
+- `Range\DateTimeRange::precision()` method.
+- `Range\TimeRange::precision()` method.
 
 ## [0.2.4](https://github.com/frostealth/kronika/releases/tag/0.2.4) – 2025-11-12
 ### Added